Fixed basic functionality
Now the basic functionality works. Users can add new
projects or new tasks. Moreover, when adding a new
task they can assign it to an existing project from
a dropdown menu or they can enter a new value, in
which case this is added to the project list and displayed.

Todo: fix UI and clean up code!